tx process logic too many panic in error handling.
Solution:
- Refactored the tx logic code from protocol processing code.
- Properly handle the errors with Result.
- Removed unnecessary traits RequestWithTx / ResponseWithCodeAndLog
