+++ chain-abci/src/app/slash_accounts.rs
@@ -77,7 +77,8 @@ impl<T: EnclaveProxy> ChainNodeApp<T> {
                 account,
                 &self.uncommitted_account_root_hash,
                 &mut self.accounts,
-            );
+            )
+            .expect("update account");
             self.uncommitted_account_root_hash = new_root;
         }
